+++ src/panels/Workflow/LearningMap/LearningMap.scss
