+++ src/panels/Workflow/LearningMap/__Points/LearningMap__Points.js
