+++ src/panels/Workflow/LearningMap/__Row/LearningMap__Row.js
