+++ src/panels/Workflow/LearningMap/__Separator/LearningMap__Separator.scss
